Projeto done

+++ Backend/src/main/java/com/example/biblioteca/livro/AutorService.java
@@ -1,11 +1,14 @@
 package com.example.biblioteca.livro;
 
+
 import com.example.biblioteca.exceptions.ResourceNotFoundException;
 import com.example.biblioteca.livro.entity.Autor;
 
 import com.example.biblioteca.livro.repository.AutorRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.GetMapping;
+
 import java.util.List;
 
 @Service
@@ -18,7 +21,7 @@ public class AutorService {
         return autorRepository.save(autor);
     }
 
-    public List<Autor> listarCategoria() {
+    public List<Autor> listarAutor() {
         return autorRepository.findAll();
     }
 
@@ -29,9 +32,22 @@ public class AutorService {
         }
         return autor.get();
     }
+   public Autor atualizar(Autor autor) throws Exception{
+        if(!autorRepository.existsById(autor.getId())) {
+            throw new Exception("Autor não encontrado!");
+        }
+        return autorRepository.save(autor);
+   }
 
-    public void remover(Integer id) {
-        consultar(id);
+   public void remover(Integer id) throws Exception {
+        if (!autorRepository.existsById(id)) {
+            throw new Exception("Autor Não encontrado");
+        }
         autorRepository.deleteById(id);
-    }
+   }
+
+//    public void remover(Integer id) {
+//        consultar(id);
+//        autorRepository.deleteById(id);
+//    }
 }
